Implementation of transferring players via sockets
This replaces the prior mechanism which used PluginMessagingChannels. A
transferring player now has their equipment/etc transferred first to the
target server, then after everything is ok a request is made to
bungeecord to actually transfer the player.

This depends on the Socket4MC plugin (formerly SocketAPI). Note that
servers still can't send messages to eachother - they send messages to
Bungee, and bungee forwards them to the destination server.

Signed-off-by: Byron Marohn <combustible@live.com>