Major Persistent changes
During my multiple runs of Whitecliffe to farm the miniboss for T5, I noticed that the mobs there with passives and abilities, the Mage, the Brute, and the rabid wolves, do not despawn at all.
This caused for a buildup of elite mobs that never despawn when you go away far enough, causing for salty deaths due to the amount of boss mobs there are.

The changes made here removed the automatic .setRemoveWhenFarAway from constructing the bossses, and changed several bosses to have this line instead. However, bosses such as PulseLaser bosses will not have it unless they have one of the boss tags that have persistence disabled, such as the GenericBoss tag.
And for the sake of Nightmare, all large aura elites will have persistence disabled