Improved raycasting using a new collidesWithSolid method.
collidesWithSolid has mappings of hitboxes of all semisolids, for
example doors or stairs. The cast ray will use this method to check if
it should break or not, and will do this with pixel precision for
semisolids. Rays will now go over slabs, through the empty corners of
stairs, and will even count the amount of bites to accurately cast
through a cake block.