Infusion balance tweaks (#151)
* Change infusion enchants to use finals for scaling

This puts the Percent-per-Level stats in an easy to modify place at the top.

* Suggested balance tweaks.

Discussion with some of the loot balance people lead to the suggestion of Tenacity being reduced to 0.5% per level and Health scaling being changed to additive scaling. These are tentative and easy to change later.

* Reverting the Vitality scaling

It was weak enough as it is. This can be changed back if it's actually too strong somehow.