Curse of Vanishing fix 2: electric boogaloo (#178)
* Handle CoV items identically to other items after shattering

Migrated the death item handling code to a function and called it from both the CoV section and the normal item section so that they're handled identically after shattering CoV items.

* Make CoV items graveable

Add Shatter_Now to the list of possible itemdeathresults for graving

* Updated more code relating to CoV and graves

How many places do we need to check the ItemDeathResult? This code seems redundant when we already check the ItemDeathResult to give the item the tags.