Balance and bug fixes for varcosa (#380)
Made the cannons particles easier to see, adjusted the health on all 3 phases, hopefully fixed the phase 2 mobs not spawning - I think the problem was that the radius of mob spawning was to small and it didnt have enough valid spots to spawn so I pumped that up. Fixed the glowing on phase 3 by reapplying the tag that the lantern looks for to give glowing. Also balanced the SpellActions to be more agressive. Also fixed a bug to make the hook boss actually work.