Firmament and Multitool fixes (#442)
* Tweaked Multitool

Changed multitool to not strip logs or path(exile, grab a basic land and put it on the battlefield tapped) grass, as well as allowed it to change when looking at pumpkins

* Fixed Firmament critical bug

The first change from the top is to make sure prismarine works no matter where the enabled thing is. Second change is to make sure it doesn't add it on if it is the last line and it hasn't already found it.