Fix Boss Block Break Cheese (#3177)
Fix oversight with blockbreak ability, especially when mFootLevelBreak is set to false.

Block break counts the number of "bad blocks" around the mob. Main issue is when it doesn't consider floor blocks as bad blocks, players are able to cheese bosses by placing a block on their head (and since the check requires more than 2 "bad blocks", it doesn't execute and the mob is stuck and cheesable)

This PR hopes to fix this by making the floor blocks be counted even if mFootBreakBlocks is set to false. It will not break the floor blocks, but it will count towards the requirement for the block break to execute, meaning it should happen more often now.

In theory:

Footlevel = 1 score
Other = 2 score

If we have a situation where mob is surrounded by 4 blocks on foot level, then it will result in 4 * 1 = 4 score.

If we check for badscore >= 6, then it will be the case that any block after on above ground (such as the boss' ceiling) will definitely cause it to execute block break.

Meaning, it will not be cheesed.