Temporarily disable training dummy hologram code (#3624)
This code causes server crashes in 1.19 during unload. There is some problem with the paper chunk rewrite for 1.19 that makes it invalid to remove entities in this way. Seems like the upstream tracking bug is https://github.com/PaperMC/Paper/issues/6451. 

Need to rewrite this code, since it looks like that upstream problem will not be fixed soon.