Ashes of Eternity rework (#4601)
Applies the effect of Resurrection upon death while holding in mainhand. When a player triggers the effect, the enchant is marked as "Used" and becomes unusable to them until the next daily reset. The enchant is refreshed and mark is removed when clicking with the item in mainhand, or on weekly update.