Improve EntityUtils.isHostile (#4676)
Notably, the check for wolf ownership was inverted, and wolves and rabbits with the Hostile tag would not automatically count as hostile.

Additionally, some small optimizations by reducing the number of tag checks and instanceof checks. Note that HostileBoss already made it so mobs with boss_hostile would always have Hostile as well.