"Owned" Infusion for Players and Guilds (#5720)
Adds an infusion, "Owned", that allows players to mark ownership on items. There are two "methods" of doing this:

    Player Ownership: Allows a player to mark an item as theirs.
    Guild Ownership: Allows a member of a guild with a specific (new) permission to mark an item as their guild's.

Adds a new command, /itemownership, that manages the infusing process. This command should, for the most part, be run through a NPC.

    /itemownership guild allows specifically mods to add any guild ownership to an item in their mainhand.
    /itemownership player/playerguild/remove, are ran through an NPC and run permission checks.

Adds a new guild permission, "guild_owned_infusion", that dictates whether a guild member is able to add the Owned infusion on items for their guild or not.
Also removes old code related to the guildplots transfer